delete items from there cart
While giving the users the ability to remove items from there cart, I stumbled across a issue with sessions. Sessions only allows one user at a time to login. So if you are logged in and a other user logs in you are logged out. I am going to further investigate this issue and try to come up with a solution. It shouldn't be complicated.