Fixed dropdown
Noticed that with previous dropdown code, when clicking on the dropdown, 50 states would be added, if clicking again then another 50 is added. This leads to like 200 states existing in the drop down. I used a return to avoid multiple insertions of states. This worked. I also notice that the rest api had random names of places that weren't part of the USA, like Ontario and Ramey, so I implemented code in the frontend to remove these irrelevant names. Overall everything works fine. I just have to adjust the validation.