Implemented username validation
In the backend I implemented a query to check if a username has been taken. If so it sends the data to the frontend allowing the validation to check if the username has been taken. If it is taken the frontend will not allow the user to create a account until they user a different username. I used a GET method in the backend.