Improving Location information of user
Made a new ejs file called UpdateUserLocation which keeps track of user location info and allows the user to update there info. I also modified the create account in the index js file so that when creating a new account a empty location form is available to be filled. So far users can only have one location in the application. Added another js file for validating the location info. Will be adding more logic to it. In the future user will have the opportunity to add multiple locations and choose which one they want there product to be shipped to